requirements.txt
google-adk==1.2.1
litellm==1.72.2
python-dotenv